feat: add foxFarming.ts resolvers module
